Add own module of token manipulation and refactor in routes with auth.
