+++ controllers/users/index.js
@@ -48,30 +48,8 @@ const input = {
   }
 }
 
-const result = {
-  /**
-   * The middleware change the request's body, transforming it until it has the json that we
-   * need to send to the client.
-   * @param {Object} req - Request object.
-   * @param {Object} res - Response object.
-   * @param {Function} next - Next function, useful to call the next middleware.
-   */
-  send: (req, res, next) => {
-    if (!req.body) {
-      const err = new Error('There is no body to send to the client.')
-      err.status = 500
-      logger.error(err)
-      // Not send this message to the client, set to null and the error middleware will put default message.
-      err.message = null
-      return next(err)
-    }
-    res.json(req.body)
-  }
-}
-
 module.exports = {
   input,
   users,
-  result,
   hashPassword
 }
