+++ routes/auth/index.js
@@ -1,16 +1,18 @@
 const router = require('express').Router()
-const {input, users, result} = require('../../controllers/auth')
+const {input, token, users} = require('../../controllers/auth')
+const {result} = require('../../controllers/utils')
 
 /* Login and creation of the user's token */
 router.post('/login',
   input.validate.login,
   users.login,
-  users.token.create,
+  token.create,
   result.send
 )
 
+/* Validate a token using get method */
 router.get('/:token',
-  users.token.validate,
+  token.validate,
   result.send
 )
