+++ routes/users/index.js
@@ -1,11 +1,13 @@
 const router = require('express').Router()
 const {roles} = require('../../models')
-const {input, users, result} = require('../../controllers/users')
-const {validateToken} = require('../../controllers/utils')
+const {token} = require('../../controllers/auth')
+const {input, users} = require('../../controllers/users')
+const {result} = require('../../controllers/utils')
 
 /* Creation of a user using POST method */
 router.post('/',
-  ...validateToken.createMiddleware([roles.admin]),
+  token.validate,
+  token.validate.roles([roles.admin, roles.companyAdmin]),
   input.validate.creation,
   users.create,
   result.send
@@ -13,21 +15,24 @@ router.post('/',
 
 /* Get information of the user using GET method */
 router.get('/:email',
-  ...validateToken.createMiddleware([roles.admin]),
+  token.validate,
+  token.validate.roles([roles.admin, roles.companyAdmin]),
   users.get,
   result.send
 )
 
 /* Edit user using PUT method */
 router.put('/:email',
-  ...validateToken.createMiddleware([roles.admin]),
+  token.validate,
+  token.validate.roles([roles.admin, roles.companyAdmin]),
   users.update,
   result.send
 )
 
 /* Remove an user with the DELETE method */
 router.delete('/:email',
-  ...validateToken.createMiddleware([roles.admin]),
+  token.validate,
+  token.validate.roles([roles.admin, roles.companyAdmin]),
   users.remove,
   result.send
 )
